Move mocked state to a separate file
